requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile '.\requirements.in'
#
asgiref==3.7.2
    # via django
build==1.0.3
    # via pip-tools
click==8.1.7
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
django==4.2.6
    # via
    #   -r .\requirements.in
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
django-filter==23.3
    # via -r .\requirements.in
djangorestframework==3.14.0
    # via
    #   -r .\requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.0
    # via -r .\requirements.in
packaging==23.2
    # via build
pip-tools==7.3.0
    # via -r .\requirements.in
pyjwt==2.8.0
    # via djangorestframework-simplejwt
pyproject-hooks==1.0.0
    # via build
pytz==2023.3.post1
    # via djangorestframework
sqlparse==0.4.4
    # via django
tzdata==2023.3
    # via django
wheel==0.41.2
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
